Updates to main, mainTests, and commands

+++ Api/main.py
@@ -1,32 +1,26 @@
-# pip install 'fastapi[standard]'
-# pip install docxtpl
-# pip install docx2pdf
-# pip install python-docx
-# pip install requests
-# pip install pillow
-
-# Used for API calls
+# Brugt til API kald
 from fastapi import FastAPI, Form, File, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import FileResponse
 import uvicorn
 
-# Used in conversion of required data, when converting from UploadFile to other type
+# Brugt til i konversionen af nødvændig data, ved konvertering fra .json til given datatype
 import json
 import shutil
 
-# Used for inserting data in template
+# Brugt til at indsætte dynamiske data i word skabelon
 from docxtpl import DocxTemplate, InlineImage
 from docx import Document
 from docx.shared import Inches
 
-# Used for converting docx-format to pdf-format
+# Brugt til at konvertere fra docx-format til pdf-format
 import os
 from docx2pdf import convert
 
-# Used for finding a given image from URL
+# Brugt til at finde et givent billede, fra URL
 import requests
 
+## --------------------------------------- Skrevet af Bjørn ----------------------------------------- ##
 app = FastAPI()
 
 app.add_middleware(
@@ -36,63 +30,75 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"]
 )
+## -------------------------------------------------------------------------------------------------- ##
 
+## ----------------------------------- Skrevet af Magnus og Bjørn ----------------------------------- ##
 @app.post('/insert-dynamic-data/')
 async def insert_dynamic_data(
         templateFile: UploadFile = File(...),
         contextFile: UploadFile = File(...),
         imageURLFile: UploadFile = File(None),
-        imagesSize: int = Form(5),
         pdfName: str = Form("Invoice"),
     ):
-    # Extracts the context-dictionary from .json file
+    ## ----------------------------------- Skrevet af Magnus ---------------------------------------- ##
+    # Ekstrahere context-dictionary og imageURLs-liste fra .json filer
     context: dict = json.loads(contextFile.file.read())
-    if imageURLFile != None: imageURLs: list = json.loads(imageURLFile.file.read())
+    if imageURLFile != None: imageURLs: dict = json.loads(imageURLFile.file.read())
     else: imageURLs = None
+    ## ---------------------------------------------------------------------------------------------- ##
 
-    # Save a copy pf the uploaded template file
+    ## ------------------------------------ Skrevet af Bjørn ---------------------------------------- ##
+    # Gemmer kopi af uploadede skabelon fil
     templatePath = 'uploadedTemplate.docx'
     with open(templatePath, "wb") as buffer:
         shutil.copyfileobj(templateFile.file, buffer)
+    ## ---------------------------------------------------------------------------------------------- ##
     
+    ## ------------------------------------ Skrevet af Magnus --------------------------------------- ##
+    # Finder og åbner den kopierede skabelon som DocxTemplate objekt
     tpl = DocxTemplate(templatePath)
     
+    # Finder givene billeder fra URL, og ligger den i context-dictionary
     if imageURLs != None:
         index = 0
         images = []
         for url in imageURLs:
             image = FindImage(url, f'image{index}')
             
-            foundImage = InlineImage(tpl, image, width = Inches(imagesSize))
+            context[f'image{index}'] = InlineImage(tpl, image, width = Inches(imageURLs[url]))
             
             images.append(image)
-            context[f'image{index}'] = foundImage
-            
             index += 1
     
-    context = InsertPageNumbers(context, templatePath)
-    
+    # Validere context- og skabelon-variabler
     errMsg, valid = ValidateVariables(templatePath, context)
     if valid == False:
         os.remove(templatePath)
         if imageURLs != None: RemoveRenderedImages(images)
         return errMsg
 
-    # Inserts data from the context-dictionary to the template
+    # Indsætter data fra context-dictionary til skabelon
     tpl.render(context)
     
+    # Gemmer og konvertere til PDF
     tpl.save(f'{pdfName}.docx')
-
     if imageURLs != None: RemoveRenderedImages(images)
     return ConvertDocxToPDF(pdfName, templatePath)
+    ## ---------------------------------------------------------------------------------------------- ##
 
-def ConvertDocxToPDF(path: str, templatePath):
+## ----------------------------------- Skrevet af Magnus og Bjørn ----------------------------------- ##
+def ConvertDocxToPDF(path: str, templatePath: str, isTest: bool = False):
+    ## ------------------------------------ Skrevet af Magnus --------------------------------------- ##
     convert(f'{path}.docx', f'{path}.pdf')
     os.remove(f'{path}.docx')
-    os.remove(templatePath)
+    if isTest == False: os.remove(templatePath)
+    ## ---------------------------------------------------------------------------------------------- ##
 
+    ## ------------------------------------ Skrevet af Bjørn ---------------------------------------- ##
     return FileResponse(f'{path}.pdf', media_type = 'application/pdf', filename = path)
+    ## ---------------------------------------------------------------------------------------------- ##
 
+## ---------------------------------------- Skrevet af Magnus --------------------------------------- ##
 def ValidateVariables(path: str, context: dict):
     doc = Document(path)
     valid = True
@@ -104,7 +110,8 @@ def ValidateVariables(path: str, context: dict):
     valuesInputted = []
 
     for p in doc.paragraphs:
-        # All variables must begin with "{{", therefore we can find them simply by looking for this
+        # Alle variabler i skabelonet starter med "{{"
+        # Derfor kan man finde alle variabler ved at lede gennem alt tekst i dokumentet efter disse
         if p.text.__contains__('{{') and not (p.text.__contains__('{% for') or p.text.__contains__('{% endfor')):
             tempValues = p.text.rsplit('{{')
 
@@ -112,20 +119,17 @@ def ValidateVariables(path: str, context: dict):
                 if not temp.__contains__('}}'):
                     tempValues.remove(temp)
 
-            # I don't know why, but it only seems to be able to do one of the variables per for-loop...
-            # It doesn't work correctly with a single for-loop, there must be one for each variable in the tempValues list
-            for _ in range(len(tempValues)):
-                for tempValue in tempValues:
-                    tempValues.remove(tempValue)
-                    tempValue = tempValue.split('}}')[0]
-                    tempValues.append(tempValue)
+            # Isolere skabelon-variablerne fra de sidste klammer, så de kan sammenlignes med context-variablerne
+            foundValues = []
+            for i in range(len(tempValues)):
+                tempValue = tempValues[i].split('}}')[0]
+                foundValues.append(tempValue)
 
-            for temp in tempValues:
-                if not values.__contains__(temp):
-                    values.append(temp)
+            for value in foundValues:
+                if not values.__contains__(value):
+                    values.append(value)
 
-    # Ignores lists, because these are not acounted for as variables in the template
-    # It is also not seen as an error, if these are empty
+    # Ignoere lister, da disse godt må være tomme
     for key in context:
         if not isinstance(context[key], list): keysNotContained.append(key)
     
@@ -141,7 +145,7 @@ def ValidateVariables(path: str, context: dict):
                 valuesInputted.append(value)
                 valuesNotInputted.remove(value)
     
-    # Builds the exception string, so it can be returned
+    # Bygger exception-string, så den kan returneres til brugeren
     errorMsg = ''
     if not len(keysNotContained) == 0:
         for key in keysNotContained:
@@ -155,23 +159,6 @@ def ValidateVariables(path: str, context: dict):
     
     return errorMsg, valid
 
-def InsertPageNumbers(context: dict, path: str):
-    doc = Document(path)
-
-    pageNumber = 1
-    for p in doc.paragraphs:
-        if p.text.__contains__('{{Current_Page}}'):
-            subStr = p.text
-            subStr = subStr.split('{{Current_Page}}')
-            subStr = subStr[0] + str(pageNumber) + subStr[1]
-            p.text = subStr
-            pageNumber += 1
-    
-    doc.save(path)
-
-    context['Total_Pages'] = str(pageNumber - 1)
-    return context
-
 def FindImage(url: str, fileName: str):
     data = requests.get(url).content
     f = open(f'{fileName}.png','wb')
@@ -183,10 +170,9 @@ def FindImage(url: str, fileName: str):
 def RemoveRenderedImages(images):
     for image in images:
         os.remove(image)
+## -------------------------------------------------------------------------------------------------- ##
 
-
-# ------------------------------------------------------------------------------------------- #
-
-
+## --------------------------------------- Skrevet af Bjørn ----------------------------------------- ##
 if __name__ == '__main__':
     uvicorn.run(app, host = '127.0.0.1', port = 8000)
+## -------------------------------------------------------------------------------------------------- ##
